latest

+++ src/main/java/com/example/umc4_heron_template_jpa/board/photo/PostPhotoRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface PostPhotoRepository  extends JpaRepository<PostPhoto, Integer> {
+public interface PostPhotoRepository  extends JpaRepository<PostPhoto, Long> {
 
 
     @Query("select p from PostPhoto p where p.board.boardId = :boardId")
